+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/Component/Collision/Algolithm/CollisionAlgolithmImplementation.cpp
@@ -6,7 +6,7 @@ std::string my::CollisionAlgolithmType::kPlayerEnemyBulletCollisionAlgolithm = "
 std::string my::CollisionAlgolithmType::kPlayerShipCollisionAlgolithm = "PlayerShipCollisionAlgolithm";
 std::string my::CollisionAlgolithmType::kPlayerWaterFlowCollisionAlgolithm = "PlayerWaterFlowCollisionAlgolithm";
 std::string my::CollisionAlgolithmType::kEnemyPlayerCollisionAlgolithm = "EnemyPlayerCollisionAlgolithm";
-std::string my::CollisionAlgolithmType::kEnemyPlayerMeleeAttackCollisionAlgolithm = "EnemyPlayerMeleeAttackCollisionAlgolithm";
+std::string my::CollisionAlgolithmType::kEnemyOmniWrenchAttackCollisionAlgolithm = "EnemyOmniWrenchAttackCollisionAlgolithm ";
 std::string my::CollisionAlgolithmType::kSightPlayerCollisionAlgolithm = "SightPlayerCollisionAlgolithm";
 std::string my::CollisionAlgolithmType::kSightEnemyCollisionAlgolithm = "SightEnemyCollisionAlgolithm";
 std::string my::CollisionAlgolithmType::kEnemyEnemyCollisionAlgolithm = "kEnemyEnemyCollisionAlgolithm";
