+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/RachetApplication.vcxproj
@@ -247,7 +247,6 @@
     <ClCompile Include="Actor\Weapon\OmniWrench.cpp" />
     <ClCompile Include="Actor\Weapon\Pyrocitor.cpp" />
     <ClCompile Include="Actor\Weapon\Weapon.cpp" />
-    <ClCompile Include="BridgeEvent.cpp" />
     <ClCompile Include="Camera\AutoCameraController.cpp" />
     <ClCompile Include="Camera\Camera.cpp" />
     <ClCompile Include="Camera\CameraController.cpp" />
@@ -351,6 +350,9 @@
     <ClCompile Include="Component\Weapon\OmniWrench\OmniWrenchActionStateComponent.cpp" />
     <ClCompile Include="Component\Weapon\OmniWrench\OmniWrenchDefaultComponent.cpp" />
     <ClCompile Include="Component\Weapon\OmniWrench\OmniWrenchThrowedComponent.cpp" />
+    <ClCompile Include="Event\BridgeEvent.cpp" />
+    <ClCompile Include="Event\ShipEvent.cpp" />
+    <ClCompile Include="Event\StageViewEvent.cpp" />
     <ClCompile Include="Factory\ActorFactory.cpp" />
     <ClCompile Include="Factory\BehaviourExecutorFactory.cpp" />
     <ClCompile Include="Factory\BehaviourFactory.cpp" />
@@ -379,7 +381,6 @@
     <ClCompile Include="Scene\SceneImplementation.cpp" />
     <ClCompile Include="Scene\SceneManager.cpp" />
     <ClCompile Include="Scene\TitleScene.cpp" />
-    <ClCompile Include="ShipEvent.cpp" />
     <ClCompile Include="Stage\Gimmick\Bridge.cpp" />
     <ClCompile Include="Stage\Gimmick\Elevator.cpp" />
     <ClCompile Include="Stage\Gimmick\GimmickBase.cpp" />
@@ -497,7 +498,6 @@
     <ClInclude Include="Behaviour\Node\SequencerNode.h" />
     <ClInclude Include="Behaviour\Node\SimplexNode.h" />
     <ClInclude Include="BezierCurveAnimation.h" />
-    <ClInclude Include="BridgeEvent.h" />
     <ClInclude Include="Camera\AutoCameraController.h" />
     <ClInclude Include="Camera\Camera.h" />
     <ClInclude Include="Camera\CameraController.h" />
@@ -606,6 +606,9 @@
     <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchComponentInclude.h" />
     <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchDefaultComponent.h" />
     <ClInclude Include="Component\Weapon\OmniWrench\OmniWrenchThrowedComponent.h" />
+    <ClInclude Include="Event\BridgeEvent.h" />
+    <ClInclude Include="Event\ShipEvent.h" />
+    <ClInclude Include="Event\StageViewEvent.h" />
     <ClInclude Include="Factory\ActorFactory.h" />
     <ClInclude Include="Factory\BehaviourExecutorFactory.h" />
     <ClInclude Include="Factory\BehaviourFactory.h" />
@@ -645,7 +648,6 @@
     <ClInclude Include="Scene\SceneDefine.h" />
     <ClInclude Include="Scene\SceneManager.h" />
     <ClInclude Include="Scene\TitleScene.h" />
-    <ClInclude Include="ShipEvent.h" />
     <ClInclude Include="Stage\Gimmick\Bridge.h" />
     <ClInclude Include="Stage\Gimmick\Elevator.h" />
     <ClInclude Include="Stage\Gimmick\GimmickBase.h" />
